requirements.txt
Django>=4.2,<5.0
python-dateutil>=2.8.2
pandas>=2.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
requests>=2.28.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

